feat: added RatingBar component and implementation

+++ app/src/main/java/com/dalbitresb/partysoul/models/Joke.kt
@@ -9,7 +9,7 @@ data class Joke(
     @ColumnInfo(name = "source") var source: String,
     @ColumnInfo(name = "joke_id") var jokeId: String?,
     @ColumnInfo(name = "content") var content: String,
-    @ColumnInfo(name = "rating") var rating: Int,
+    @ColumnInfo(name = "rating") var rating: Float,
     @ColumnInfo(name = "created_at") var createdAt: String,
 ) {
     @PrimaryKey(autoGenerate = true)
